refactor: address review findings — remove non-null assertions and extract shared schema
- Replace 3 `customRoleId!` non-null assertions with type predicates
  and `.filter().map()` patterns per project rules
- Extract `teamRoleInputSchema` to `routers/schemas/teamRole.ts` to
  break the circular import from invite.ts → organization.ts
- Update both routers to import from the shared location
- Remove unused `TeamUserRole` import from organization.ts
